+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/InsecureCookieBasedOAuthPersistence.kt
@@ -6,6 +6,7 @@ import org.http4k.core.Status.Companion.FORBIDDEN
 import org.http4k.core.cookie.Cookie
 import org.http4k.core.cookie.cookie
 import org.http4k.core.cookie.invalidateCookie
+import org.http4k.security.openid.Nonce
 import java.time.Clock
 import java.time.Duration
 import java.time.LocalDateTime
@@ -21,17 +22,28 @@ class InsecureCookieBasedOAuthPersistence(cookieNamePrefix: String,
 
     private val csrfName = "${cookieNamePrefix}Csrf"
 
+    private val nonceName = "${cookieNamePrefix}Nonce"
+
     private val accessTokenCookieName = "${cookieNamePrefix}AccessToken"
 
     override fun retrieveCsrf(request: Request) = request.cookie(csrfName)?.value?.let(::CrossSiteRequestForgeryToken)
 
     override fun retrieveToken(request: Request): AccessToken? = request.cookie(accessTokenCookieName)?.value?.let { AccessToken(it) }
 
+    override fun retrieveNonce(request: Request): Nonce? = request.cookie(nonceName)?.value?.let { Nonce(it) }
+
     override fun assignCsrf(redirect: Response, csrf: CrossSiteRequestForgeryToken) = redirect.cookie(expiring(csrfName, csrf.value))
 
-    override fun assignToken(request: Request, redirect: Response, accessToken: AccessToken) = redirect.cookie(expiring(accessTokenCookieName, accessToken.value)).invalidateCookie(csrfName)
+    override fun assignToken(request: Request, redirect: Response, accessToken: AccessToken) = redirect.cookie(expiring(accessTokenCookieName, accessToken.value))
+        .invalidateCookie(csrfName)
+        .invalidateCookie(nonceName)
 
-    override fun authFailureResponse() = Response(FORBIDDEN).invalidateCookie(csrfName).invalidateCookie(accessTokenCookieName)
+    override fun assignNonce(redirect: Response, nonce: Nonce): Response = redirect.cookie(expiring(nonceName, nonce.value))
+
+    override fun authFailureResponse() = Response(FORBIDDEN)
+        .invalidateCookie(csrfName)
+        .invalidateCookie(accessTokenCookieName)
+        .invalidateCookie(nonceName)
 
     private fun expiring(name: String, value: String) = Cookie(name, value, expires = LocalDateTime.ofInstant(clock.instant().plus(cookieValidity), ZoneId.of("GMT")))
 }