+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/OAuthCallback.kt
@@ -21,18 +21,28 @@ class OAuthCallback(
                 ?.let(::CrossSiteRequestForgeryToken)
                 ?.takeIf { it == oAuthPersistence.retrieveCsrf(request) }
                 ?.let {
-                    request.queryOrFragmentParameter("id_token")?.let { idTokenConsumer.consumeFromAuthorizationResponse(IdToken(it)) }
-                    accessTokenFetcher.fetch(code)
-                        ?.let { tokenDetails ->
-                            tokenDetails.idToken?.also(idTokenConsumer::consumeFromAccessTokenResponse)
-
-                            val originalUri = state.find { it.first == "uri" }?.second ?: "/"
-                            oAuthPersistence.assignToken(request, Response(TEMPORARY_REDIRECT)
-                                .header("Location", originalUri), tokenDetails.accessToken)
-                        }
+                    val idToken = request.queryOrFragmentParameter("id_token")?.let { IdToken(it) }
+                    if(hasValidNonceInIdToken(request, idToken)) {
+                        idToken?.let {  idTokenConsumer.consumeFromAuthorizationResponse(it)}
+                        accessTokenFetcher.fetch(code)
+                            ?.let { tokenDetails ->
+                                tokenDetails.idToken?.also(idTokenConsumer::consumeFromAccessTokenResponse)
+                                val originalUri = state.find { it.first == "uri" }?.second ?: "/"
+                                oAuthPersistence.assignToken(request, Response(TEMPORARY_REDIRECT)
+                                    .header("Location", originalUri), tokenDetails.accessToken)
+                            }
+                    } else {
+                        null
+                    }
                 }
         }
         ?: oAuthPersistence.authFailureResponse()
+
+    private fun hasValidNonceInIdToken(request: Request, idToken: IdToken?): Boolean {
+        return if(idToken != null) {
+            idTokenConsumer.nonceFromIdToken(idToken) == oAuthPersistence.retrieveNonce(request)
+        } else true
+    }
 }
 
 private fun Request.queryOrFragmentParameter(name: String): String? = this.query(name) ?: fragmentParameter(name)