+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/AuthRequest.kt
@@ -3,6 +3,7 @@ package org.http4k.security.oauth.server
 import org.http4k.core.Uri
 import org.http4k.security.ResponseType
 import org.http4k.security.ResponseType.Code
+import org.http4k.security.openid.Nonce
 import org.http4k.security.openid.RequestJwtContainer
 import org.http4k.security.openid.RequestJwts
 
@@ -12,7 +13,7 @@ data class AuthRequest(
     val redirectUri: Uri,
     val state: String?,
     val responseType: ResponseType = Code,
-    val nonce: String? = null,
+    val nonce: Nonce? = null,
     val request: RequestJwtContainer? = null) {
 
     fun isOIDC() = scopes.map { it.toLowerCase() }.contains(OIDC_SCOPE)