+++ http4k-security-oauth/src/test/kotlin/org/http4k/security/oauth/server/AuthenticationCompleteTest.kt
@@ -53,7 +53,7 @@ class AuthenticationCompleteTest {
             and hasHeader("location",
             authorizationRequest.redirectUri
                 .fragmentParameter("code", "dummy-token-for-jdoe")
-                .fragmentParameter("id_token", "dummy-id-token-for-jdoe")
+                .fragmentParameter("id_token", "dummy-id-token-for-jdoe-nonce:unknown")
                 .fragmentParameter("state", "some state").toString()))
     }